fixed gaussian code

+++ src/thread.h
@@ -1,3 +1,3 @@
 #include <thread>
 
-void work (const int, const int, const int, const int, const int, const int, const double, const int); 
+void work (const int, const int, const int, const int, const int, const int, const double, const double);
